Refactor markdown editor
