+++ src/editor/containers/MarkdownEditor.tsx
@@ -1,6 +1,6 @@
-import React, { useCallback } from 'react'
+import React, { useCallback, useEffect, useMemo, useRef, useState } from 'react'
 
-import { stripVisualIndents } from '../utils/visualIndent'
+import { injectVisualIndents, stripVisualIndents } from '../utils/visualIndent'
 import { computeListEnter } from '../utils/computeListEnter'
 
 import { useDispatch } from 'react-redux'
@@ -13,14 +13,24 @@ import { HotkeyId } from '../../hotkeys/registry'
 import MarkdownEditorComponent from '../components/MarkdownEditor'
 
 const MarkdownEditor: React.FC = () => {
+  const [charsPerLine, setCharsPerLine] = useState<number>(80)
+
+  const containerRef = useRef<HTMLDivElement>(null)
+
   const content = useEditorText()
   const captureTab = useCaptureTabEnabled()
 
   const dispatch = useDispatch()
 
+  const displayContent = useMemo(
+    () => injectVisualIndents(content, charsPerLine),
+    [content, charsPerLine],
+  )
+
   const handleContentChange = useCallback(
     (content: string) => {
-      dispatch(setText(content))
+      const sanitized = stripVisualIndents(content)
+      dispatch(setText(sanitized))
     },
     [dispatch],
   )
@@ -60,7 +70,7 @@ const MarkdownEditor: React.FC = () => {
 
       e.preventDefault()
 
-      handleContentChange(stripVisualIndents(newValue))
+      handleContentChange(newValue)
 
       // Restore caret position in the next tick once React has flushed the
       // state update and the value prop has propagated to the underlying
@@ -72,12 +82,90 @@ const MarkdownEditor: React.FC = () => {
     [handleContentChange],
   )
 
+  const focusEditor = useCallback(() => {
+    if (containerRef.current) {
+      const textarea = containerRef.current.querySelector('textarea')
+      if (textarea) {
+        ;(textarea as HTMLTextAreaElement).focus()
+
+        // Move cursor to the end
+        const len = (textarea as HTMLTextAreaElement).value.length
+        ;(textarea as HTMLTextAreaElement).setSelectionRange(len, len)
+      }
+    }
+  }, [containerRef])
+
+  // ------------------------------------------------------------------
+  // Determine roughly how many monospace characters fit on a single
+  // physical line inside the textarea.  This allows us to mimic the
+  // browser's soft-wrapping so that we can inject matching visual
+  // indents for wrapped list items.
+  // ------------------------------------------------------------------
+  useEffect(() => {
+    const calc = () => {
+      const textarea = containerRef.current?.querySelector('textarea') as
+        | HTMLTextAreaElement
+        | undefined
+      if (!textarea) return
+
+      const cs = window.getComputedStyle(textarea)
+
+      const padLeft = parseFloat(cs.paddingLeft || '0')
+      const padRight = parseFloat(cs.paddingRight || '0')
+      const availWidth = textarea.clientWidth - padLeft - padRight
+
+      // Measure width of a single glyph
+      const m = document.createElement('span')
+      m.textContent = 'a'
+      m.style.fontFamily = cs.fontFamily
+      m.style.fontSize = cs.fontSize
+      m.style.whiteSpace = 'pre'
+      m.style.visibility = 'hidden'
+      document.body.appendChild(m)
+      const charWidth = m.getBoundingClientRect().width || 8
+      document.body.removeChild(m)
+
+      if (!charWidth) return
+
+      const per = Math.floor(availWidth / charWidth)
+      if (per > 0 && per !== charsPerLine) {
+        setCharsPerLine(per)
+      }
+    }
+
+    calc()
+
+    const ro = new ResizeObserver(calc)
+    const textarea = containerRef.current?.querySelector('textarea')
+    if (textarea) ro.observe(textarea)
+
+    window.addEventListener('resize', calc)
+
+    return () => {
+      ro.disconnect()
+      window.removeEventListener('resize', calc)
+    }
+  }, [charsPerLine])
+
+  // Focus the editor initially
+  useEffect(() => {
+    focusEditor()
+  }, [focusEditor])
+
+  // Focus the editor when a new file is created
+  useEffect(() => {
+    if (content === '') {
+      focusEditor()
+    }
+  }, [content, focusEditor])
+
   return (
     <MarkdownEditorComponent
-      content={content}
+      content={displayContent}
       onContentChange={handleContentChange}
       captureTab={captureTab}
       onKeyDown={handleKeyDown}
+      containerRef={containerRef}
     />
   )
 }
